+++ src/pytorch_lightning/strategies/tpu_spawn.py
@@ -97,6 +97,10 @@ class TPUSpawnStrategy(DDPSpawnStrategy):
 
         return xm.xla_device()
 
+    @property
+    def local_rank(self) -> int:
+        return self.cluster_environment.local_rank() if self.cluster_environment is not None else 0
+
     @staticmethod
     def _validate_dataloader(dataloaders: Union[TRAIN_DATALOADERS, EVAL_DATALOADERS]) -> None:
         def check_has_len(dataloader: DataLoader) -> None:
@@ -233,6 +237,11 @@ class TPUSpawnStrategy(DDPSpawnStrategy):
         self._launched = True
         self.set_world_ranks()
         rank_zero_only.rank = self.global_rank
+
+    def set_world_ranks(self) -> None:
+        if self.cluster_environment is None:
+            return
+        rank_zero_only.rank = self.cluster_environment.global_rank()
 
     def validation_step(self, *args: Any, **kwargs: Any) -> Optional[STEP_OUTPUT]:
         assert self.model is not None